⚡️refactor main.tsx, extract homepage to home.tsx
