+++ .gitpod.yml
@@ -3,13 +3,16 @@ image: nberlette/gitpod-enhanced:latest
 tasks:
   - # ensure deno is installed before proceeding
     init: |
-      export PATH="$HOME/.deno/bin:$PATH"
+      export DENO_INSTALL_ROOT="$HOME/.deno/bin"
+      export PATH="$DENO_INSTALL_ROOT:$PATH"
       which deno &>/dev/null || brew install deno &>/dev/null;
       gp sync-done listo
   - # proceed if ready
     init: gp sync-await listo
     command: |
-      deno fmt && deno lint
+      deno cache
+      deno fmt
+      deno lint
 
 github:
   prebuilds:
